Moved some lambda classes to within the stream package Added some basic stream examples with tests
