--- src/main/java/Lambdas/CoderCrunchExamples/InternalExternalIteration/InternalIteration/Person.java
@@ -1,28 +1,0 @@
-package Lambdas.CoderCrunchExamples.InternalExternalIteration.InternalIteration;
-
-public class Person {
-
-    private String name;
-    private int age;
-
-    public Person(String name, int age) {
-        this.name = name;
-        this.age = age;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-}
