--- src/main/java/Lambdas/CoderCrunchExamples/TwoLevelSorting/TestStaticMethod.java
@@ -1,8 +1,0 @@
-package Lambdas.CoderCrunchExamples.TwoLevelSorting;
-
-public interface TestStaticMethod {
-
-    static void printText(){
-        System.out.println("Here is the static method");
-    };
-}
