+++ src/main/java/Streams/Lambdas/CoderCrunchExamples/SimpleSorting/Java8Lambda/Person.java
@@ -1,0 +1,28 @@
+package Streams.Lambdas.CoderCrunchExamples.SimpleSorting.Java8Lambda;
+
+public class Person {
+
+    private String name;
+    private int age;
+
+    public Person(String name, int age) {
+        this.name = name;
+        this.age = age;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getAge() {
+        return age;
+    }
+
+    public void setAge(int age) {
+        this.age = age;
+    }
+}
