+++ src/main/java/Streams/Lambdas/CoderCrunchExamples/TwoLevelSorting/TestStaticMethod.java
@@ -1,0 +1,8 @@
+package Streams.Lambdas.CoderCrunchExamples.TwoLevelSorting;
+
+public interface TestStaticMethod {
+
+    static void printText(){
+        System.out.println("Here is the static method");
+    };
+}
